Only ignore 'temporary' objects in the ObjectAddedEvent handler.
If we ignore temporary objects always, we cannot add references when first creating an object, because that'll try and create a KeyReference too, while the object is still inside of the portal_factory tool. Ignoring the add event will avoid creating a transaction on GET, while creating an explicit KeyReference for references is only done when attempting to save the object being created through the portal_factory, leading to a commit or abort depending on other factors.

We can only test for this in the original event handler, because the event handler for IPersistent is also invoked even if we were to register a more 'specific' handler for OFS.interfaces.IZopeObject.